pyproject.toml
[project]
name = "amrita-plugin-omikuji"
version = "0.1.0"
description = "AmritaBot的御神签插件"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.3",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-localstore>=0.7.4",
    "aiofiles>=24.1.0",
    "nonebot-plugin-orm>=0.8.2",
    "amrita[full]>=0.4.4",
    "zipp>=3.23.0",
]
license = "GPL-3.0-or-later"
keywords = ['chat",  "nonebot']

[project.urls]
"Homepage" = "https://github.com/AmritaBot/amrita_plugin_omikuji"
"Source" = "https://github.com/AmritaBot/amrita_plugin_omikuji"
"Issue Tracker" = "https://github.com/AmritaBot/amrita_plugin_omikuji/issues"

[tool.setuptools.packages.find]
include = ["amrita_plugin_omikuji*"]
namespaces = false

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.8",
    "nonebot-plugin-orm[default]>=0.8.2",
    "pip>=25.2",
    "pyright>=1.1.403",
]


[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"

[tool.nonebot]
plugins = [
    "amrita.plugins.chat",
    "nonebot_plugin_localstore",
    "amrita_plugin_omikuji",
]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
